refactor: correct eslint errors
