+++ .eslintrc
@@ -31,6 +31,7 @@
     "linebreak-style": 0,
     "import/no-unresolved": [2, { "caseSensitive": false }],
     "no-underscore-dangle": "off",
-    "class-methods-use-this": "off"
+    "class-methods-use-this": "off",
+    "no-shadow": 1
   }
 }
