+++ src/utils/deepMerge.ts
@@ -1,7 +1,7 @@
 import { isObject } from './checkers and validators/isObject';
 
 function deepMerge(lhs: Indexed, rhs: Indexed): Indexed {
-  const target: Indexed = lhs ? lhs : {};
+  const target: Indexed = lhs || {};
 
   Object.entries(rhs).forEach(([key, value]) => {
     if (isObject(value) && target[key] && isObject(target[key])) {
