+++ src/utils/getUserDataArray.ts
@@ -1,8 +1,12 @@
 export const getUserDataArray = (user: UserType) => {
-  return Object.entries(user).map(([title, data]) => {
-    if (title !== 'id' && title !== 'avatar') {
-      const value = data || '';
-      return { title: title, data: value };
-    }
-  });
+  return Object.entries(user)
+    .map(([title, data]) => {
+      if (title !== 'id' && title !== 'avatar') {
+        const value = data || '';
+        return { title, data: value };
+      }
+
+      return null;
+    })
+    .filter((item) => Boolean(item));
 };
